requirements.txt
opencv-python==4.10.0.84
tqdm
tensorboard==2.14.0
tensorboardx==2.6.2.2
ptflops==0.7.2.2
einops==0.8.0
timm==1.0.9
numpy==1.24.4
matplotlib==3.7.5
pytorch-msssim
